Add tests (#12)

+++ wom/models/players/models.py
@@ -72,7 +72,7 @@ class Skill(BaseModel):
     experience: int
     """The players experience in the skill."""
 
-    ehp: int
+    ehp: float
     """The players efficient hours played for the skill."""
 
 
@@ -89,7 +89,7 @@ class Boss(BaseModel):
     kills: int
     """The number of kills the player has."""
 
-    ehb: int
+    ehb: float
     """The players efficient hours bossed for the boss."""
 
 
@@ -119,7 +119,7 @@ class ComputedMetric(BaseModel):
     rank: int
     """The players rank in the computed metric."""
 
-    value: int
+    value: float
     """The value of the computed metric."""
